backend/requirements.txt
Django==1.8.18
django-cors-headers==2.1.0
django-filter==1.0.4
djangorestframework==3.6.3
Markdown==2.6.8
olefile==0.44
Pillow==4.3.0
psycopg2==2.7.3.2
